currencyrate/requirements.txt
pyln-client>=23
requests>=2.32.2
requests[socks]>=2.32.2
urllib3>=2.2.2
cachetools>=5.3.3
